[#167] refactor : CourseListWrapper 디렉토리 변경
