+++ src/main/java/com/b6/mypaldotrip/domain/course/controller/dto/response/CourseListWrapper.java
@@ -1,0 +1,7 @@
+package com.b6.mypaldotrip.domain.course.controller.dto.response;
+
+import java.util.List;
+import lombok.Builder;
+
+@Builder
+public record CourseListWrapper(List<CourseListRes> courseListResList) {}
