+++ src/test/java/com/b6/mypaldotrip/domain/course/service/CourseServiceTest.java
@@ -15,10 +15,10 @@ import com.b6.mypaldotrip.domain.course.controller.dto.request.CourseSearchReq;
 import com.b6.mypaldotrip.domain.course.controller.dto.request.CourseUpdateReq;
 import com.b6.mypaldotrip.domain.course.controller.dto.response.CourseDeleteRes;
 import com.b6.mypaldotrip.domain.course.controller.dto.response.CourseGetRes;
+import com.b6.mypaldotrip.domain.course.controller.dto.response.CourseListWrapper;
 import com.b6.mypaldotrip.domain.course.controller.dto.response.CourseSaveRes;
 import com.b6.mypaldotrip.domain.course.controller.dto.response.CourseUpdateRes;
 import com.b6.mypaldotrip.domain.course.store.entity.CourseEntity;
-import com.b6.mypaldotrip.domain.course.store.entity.CourseListWrapper;
 import com.b6.mypaldotrip.domain.course.store.repository.CourseRepository;
 import com.b6.mypaldotrip.domain.courseFile.store.repository.CourseFileRepository;
 import com.b6.mypaldotrip.domain.trip.service.TripService;
